# target_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpecForge Target API",
    "description": "Controlled system under test: auth + glucose readings CRUD.",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new user. New accounts always get the 'user' role.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "409": {
            "description": "That email is already registered."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange email and password for a bearer token.",
        "operationId": "login_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_auth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect email or password."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Return the authenticated caller's own user record.",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/readings": {
      "post": {
        "tags": [
          "readings"
        ],
        "summary": "Create a reading owned by the caller.",
        "operationId": "create_reading_readings_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "List the caller's readings. Admins see readings from all users.",
        "operationId": "list_readings_readings_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadingOut"
                  },
                  "title": "Response List Readings Readings Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/readings/{reading_id}": {
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "Fetch one reading. Only the owner or an admin may read it.",
        "operationId": "get_reading_readings__reading_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reading_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reading Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "403": {
            "description": "The reading belongs to another user and the caller is not an admin."
          },
          "404": {
            "description": "No reading with this id exists."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "readings"
        ],
        "summary": "Replace one reading. Only the owner or an admin may update it.",
        "operationId": "update_reading_readings__reading_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reading_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reading Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "403": {
            "description": "The reading belongs to another user and the caller is not an admin."
          },
          "404": {
            "description": "No reading with this id exists."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "readings"
        ],
        "summary": "Delete one reading. Only the owner or an admin may delete it.",
        "operationId": "delete_reading_readings__reading_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "reading_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reading Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid credentials."
          },
          "403": {
            "description": "The reading belongs to another user and the caller is not an admin."
          },
          "404": {
            "description": "No reading with this id exists."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_login_auth_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_auth_token_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReadingCreate": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Measured value. The accepted range depends on `unit`: mg/dL accepts 20\u2013600 inclusive, mmol/L accepts 1.1\u201333.3 inclusive. A value outside the range for its unit is rejected with 422."
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "trend": {
            "$ref": "#/components/schemas/Trend"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 280
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "type": "object",
        "required": [
          "value",
          "unit",
          "trend"
        ],
        "title": "ReadingCreate"
      },
      "ReadingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "taken_at": {
            "type": "string",
            "format": "date-time",
            "title": "Taken At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "value",
          "unit",
          "trend",
          "note",
          "taken_at",
          "created_at"
        ],
        "title": "ReadingOut"
      },
      "ReadingUpdate": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Measured value. The accepted range depends on `unit`: mg/dL accepts 20\u2013600 inclusive, mmol/L accepts 1.1\u201333.3 inclusive. A value outside the range for its unit is rejected with 422."
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "trend": {
            "$ref": "#/components/schemas/Trend"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 280
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "type": "object",
        "required": [
          "value",
          "unit",
          "trend"
        ],
        "title": "ReadingUpdate"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
      "Trend": {
        "type": "string",
        "enum": [
          "rising",
          "falling",
          "steady"
        ],
        "title": "Trend"
      },
      "Unit": {
        "type": "string",
        "enum": [
          "mg/dL",
          "mmol/L"
        ],
        "title": "Unit"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}